% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{Ellipse Area Distribution}
\usage{
area(data, ellipse_representative = TRUE)
}
\arguments{
\item{data}{named list of 3 data frames output from the `breach()` function:
`surveyData`, `breachPositionEnsemble` and `breachPositionBestFit`.}

\item{ellipse_representative}{Calculate the ellipse based on distance to 95
percentile of resamples on each passing status on each transect (TRUE)
rather than 95 percentile of areas predicted ellipse area resamples
(FALSE). Setting this to TRUE provides an ellipse polygon that reflects the
95% area m2 exactly. Otherwise, the ellipse polygon does not represent
exactly the area in m2.}
}
\value{
Named list of 7 objects:
\describe{
\item{ellipse}{sf object of the ellipse area}
\item{fifthPercentileArea}{Named list containing numeric value in metres of
5th percentile area, package version and date}
\item{spotfire_ellipse}{TIBCO Spotfire compatible geometry of ellipse}
\item{ellipseResult}{Ellipse result either Dummy or Actual Ellipse}
\item{ellipseArea}{Area distributions}
\item{warning1}{Warning if distance to good has been established for all
transects}
\item{warning2}{Warning if insufficient transect have reached good status to
constrain ellipse}
}
}
\description{
Calculate the ellipse mixing area distribution and create a ellipse polygon.
}
\examples{
\dontrun{
probability <- probability_non_linear(demo_iqi)
breach <- breach(probability)
area <- area(breach)
plot(area[["ellipse"]])
area[["fifthPercentileArea"]]
}
}
