% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive-stations.R
\name{consecutive_stations}
\alias{consecutive_stations}
\title{Consecutive Stations}
\usage{
consecutive_stations(data, pass_fail = 0.64, method = "iqi", average = FALSE)
}
\arguments{
\item{data}{Data frame with survey data}

\item{pass_fail}{Pass-fail boundary value.}

\item{method}{Type of method used to analyse samples, either "iqi" or
"residue".}
}
\value{
A named list of two data frames `sample_point_checks` and
  `survey_data`
}
\description{
Check if two consecutive stations reaching the correct standard and minimum
number of stations have been sampled.
}
\examples{
\dontrun{
stations <- consecutive_stations(demo_iqi)
}
}
