% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-hera.R
\name{convert_hera}
\alias{convert_hera}
\title{Convert to hera}
\usage{
convert_hera(method, data, overrides, breachs, areas, ellipse_representative)
}
\arguments{
\item{method}{Type of method used to analyse samples either 'iqi' or
'residue'.}

\item{data}{dataframe returned from `consecutive_stations()`}

\item{overrides}{dataframe returned from  `override()`}

\item{breachs}{dataframe returned from  `breach()`}

\item{areas}{dataframe returned from `area()`}

\item{ellipse_representative}{Calculate the ellipse based on distance to 95
percentile of resamples on each passing status on each transect (TRUE)
rather than 95 percentile of areas predicted ellipse area resamples
(FALSE). Setting this to TRUE provides an ellipse polygon that reflects the
95% area m2 exactly. Otherwise, the ellipse polygon does not represent
exactly the area in m2.}
}
\value{
dataframe in 'hera' format
}
\description{
Convert kraken output into 'hera' format which is a long format tibble
dataframe
}
