% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-residue.R
\docType{data}
\name{demo_residue}
\alias{demo_residue}
\title{Demo residue data}
\format{
A data frame with 84 rows and 7 variables.
\describe{
\item{Survey_date}{Survey_date character string}
\item{MCFF}{MCFF Site name - character string}
\item{Transect}{Transect integer}
\item{Station}{Station integer}
\item{Easting}{Easting coordinate}
\item{Northing}{Northing coordinate}
\item{IQI}{IQI ratio - Environmental Quality Ratio EQR}
}
}
\usage{
demo_residue
}
\description{
Mocked data for illustrative purposes
}
\keyword{datasets}
