% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomechanicsFunctions.R
\name{interpolateR}
\alias{interpolateR}
\title{Interpolate numerical data}
\usage{
interpolateR(df, n, method = 'spline', ...)
}
\arguments{
\item{\code{df}}{Numerical values that represent points to interpolate. This could either be a vector or a data.frame.}

\item{\code{n}}{An integer numver for the number of points that you want to divide your data into.}

\item{\code{method}}{A character string to indicate the interpolation method, following the description in the signal::interp1() function.The 'spline' is chosen as a default.}
}
\description{
This is a basic procedure to interpolate numerica data into a different number of points, which is partly useful to standardize data in kinematic analyses.
}
\details{
These procedures follow the methodology described in Kawano et al. 2016.
}
\examples{

x <- seq(1, 10, 1)
y <- rnorm(10, 5, 0.1)
xy <- data.frame(x,y)
interpolateR(xy, 20, method = "cubic")


}
\references{
Kawano SM, Economy DR, Kennedy MS, Dean D, Blob RW. 2016. Comparative limb bone loading in the humerus and femur of the tiger salamander Ambystoma tigrinum: testing the "mixed-chain" hypothesis for skeletal safety factors. Journal of Experimental Biology 219: 341-353. \url{http://jeb.biologists.org/content/219/3/341}
}
