% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomechanicsFunctions.R
\name{pitchAngle}
\alias{pitchAngle}
\title{Calculate pitch angle (in degrees) from XYZ coordinate data of two points}
\usage{
pitchAngle(P1, P2)
}
\arguments{
\item{\code{P1}}{A data.frame of numeric values containing the X, Y, and Z coordinate data, respectively, for the first point (e.g, the pelvic girdle)}

\item{\code{P2}}{A data.frame of numeric values containing the X, Y, and Z coordinate data, respectively, for the second point (e.g., the hip)}
}
\description{
Calculates the pitch angle (in degrees), from the XYZ coordinates of two points
}
\details{
These procedures follow the methodology used in Kawano and Blob (2013) and Kawano et al. 2016 to calculate angles formed about the limb joints in animals.
}
\examples{

P1 <- matrix(c(0.001005251,	-0.09343679,	0.01603717, 0.001195392,	-0.09347614,	0.0161097, 0.001406054,	-0.09350069,	0.01616693), 3, 3)
P2 <- matrix(c(0.006305306,	-0.08206114,	0.006997669, 0.006526961,	-0.08207707,	0.006980824, 0.006747555,	-0.08207049,	0.006975157), 3, 3)

pitch <- pitchAngle(P1, P2)

}
\references{
Kawano SM, Blob RW. 2013. Propulsive forces of mudskipper fins and salamander limbs during terrestrial locomotion: implications for the invasion of land. Integrative and Comparative Biology 53(2): 283-294. \url{https://academic.oup.com/icb/article/53/2/283/806410/Propulsive-Forces-of-Mudskipper-Fins-and}

Kawano SM, Economy DR, Kennedy MS, Dean D, Blob RW. 2016. Comparative limb bone loading in the humerus and femur of the tiger salamander Ambystoma tigrinum: testing the "mixed-chain" hypothesis for skeletal safety factors. Journal of Experimental Biology 219: 341-353. \url{http://jeb.biologists.org/content/219/3/341}
}
