% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomechanicsFunctions.R
\name{profilePlotR}
\alias{profilePlotR}
\title{Generating profile plots for longitudinal data, with repeated measures}
\usage{
profilePlotR(d = d, xname = xname, yname = yname, groupname = groupname, subgroupname = subgroupname, rowname = rowname, colors = c("red", "blue"), title = "plot", xlab = "x", ylab = "y", highlight = NULL, ...)
}
\arguments{
\item{\code{d}}{data (currently only accepts input for one variable at a time).}

\item{\code{xname}}{x-axis variable name.}

\item{\code{yname}}{y-axis variable name.}

\item{\code{groupname}}{variable name for the overall group that is being evaluated (e.g., species).}

\item{\code{subgroupname}}{variable name for a subgroup of the overall group (e.g., individual within species)}

\item{\code{rowname}}{variable name for the rows (e.g., ID number).}

\item{\code{title}}{character string for the title of the plot.}

\item{\code{xlab}}{character string for the x-axis label.}

\item{\code{ylab}}{character string for the y-axis label.}

\item{\code{colorlinesby}}{variable name for the grouping by which to color the individual lines by.}

\item{\code{highlight}}{Optional feature to highlight certain data points.}
}
\description{
Generating profile plot of multiple trials
}
\details{
Function to quickly generate profile plots for data. For instance, kinematic plots over time for multiple individuals that have multiple trials of data collected.
}
\examples{

profilePlotR(subset(AT_Kine2, Variable == "AbductAdductAngle"), "PercentStance", "value", groupname = "Appendage", subgroupname = "Ind", rowname = "Filename", highlight = AT_Kine2_AAA_subset, title = "Abduction versus Adduction", xlab = "PercentStance", ylab = "Degrees")

}
