% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomechanicsFunctions.R
\name{removeOverlaps}
\alias{removeOverlaps}
\title{Remove sections of the data where there is overlap}
\usage{
removeOverlaps <- function(df, primary, secondary, filmRate, ...)
}
\arguments{
\item{\code{df}}{A data.frame containing variables of interest.}

\item{\code{primary}}{A data.frame containing two numeric values that correspond to the beginning and end of an event (e.g., stance phase of forelimb).}

\item{\code{secondary}}{A data.frame containing two numeric values that correspond to the beginning and end of an event (e.g., stance phase of hind limb) that would affect interpretations of the primary event.}

\item{\code{filmRaate}}{A numeric value of the filming rate.}
}
\description{
Identifies portions of the dataset where two occurrences overlap and replaces with NAs.
}
\details{
These procedures follow the methodology used in Kawano and Blob (2013) and Kawano et al. 2016.
}
\examples{

primary <- data.frame(35, 148)
secondary <- data.frame(128, 244)
removeOverlaps(Pel_GRFs_Filtered_dataset, primary, secondary, 100)

}
\references{
Kawano SM, Blob RW. 2013. Propulsive forces of mudskipper fins and salamander limbs during terrestrial locomotion: implications for the invasion of land. Integrative and Comparative Biology 53(2): 283-294. \url{https://academic.oup.com/icb/article/53/2/283/806410/Propulsive-Forces-of-Mudskipper-Fins-and}

Kawano SM, Economy DR, Kennedy MS, Dean D, Blob RW. 2016. Comparative limb bone loading in the humerus and femur of the tiger salamander Ambystoma tigrinum: testing the "mixed-chain" hypothesis for skeletal safety factors. Journal of Experimental Biology 219: 341-353. \url{http://jeb.biologists.org/content/219/3/341}
}
