% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomechanicsFunctions.R
\name{smootheR}
\alias{smootheR}
\title{Smooth numerical data using a polynomial smoothing spline}
\usage{
smootheR(df, method = 1, norder = 3, spar = NULL, ...)
}
\arguments{
\item{\code{df}}{Numerical values that represent points to smooth. This could either be a vector or a data.frame.}

\item{\code{method}}{An integer numver that codes for the method is based on the different options available to smooth the data, following the description in the signal::smooth.Pspline() function.}

\item{\code{norder}}{An integer number that codes the order for the polynomial used for the spline, following the description in the signal::smooth.Pspline() function.}

\item{\code{spar}}{An optional vector that can be provided to use custom parameters for smoothing the data, following the description in the signal::smooth.Pspline() function.}
}
\description{
This is a basic procedure to smooth numerica data using splines, which is partly useful for kinematic analyses.
}
\details{
These procedures follow the methodology described in Kawano et al. 2016.
}
\examples{



}
\references{
Kawano SM, Economy DR, Kennedy MS, Dean D, Blob RW. 2016. Comparative limb bone loading in the humerus and femur of the tiger salamander Ambystoma tigrinum: testing the "mixed-chain" hypothesis for skeletal safety factors. Journal of Experimental Biology 219: 341-353. \url{http://jeb.biologists.org/content/219/3/341}
}
