% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-import.R
\name{survey_import}
\alias{survey_import}
\title{Import Survey Template}
\usage{
survey_import(path = NULL, spreadsheet_version = 2)
}
\arguments{
\item{path}{character file path to survey template}

\item{spreadsheet_version}{Version before 2024 (ish) is default '1', current version is '2'.}
}
\value{
Data frame with 5 variables
\describe{
\item{project_id}{character project id}
\item{sample_id}{character sample id}
\item{question}{question - character question}
\item{response}{response - character response value}
\item{label}{label - Mainly used for labelling taxonomic observations}
}
}
\description{
Import Survey Template
}
\examples{
\dontrun{
file <- system.file("extdat",
  "demo-data/220421-SelfMon-N4952-CAV1-Enhanced.xlsx",
  package =
    "kraken"
)
data <- survey_import(file)
}

}
