% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomechanicsFunctions.R
\name{voltToForce}
\alias{voltToForce}
\title{Convert voltage data to ground reaction force data}
\usage{
voltToForce(calib, lightStartFrame, startFrame, endFrame, videoHz, forceHz, zeroStart)
}
\arguments{
\item{\code{calib}}{A vector of numeric data containing the calibrations in the vertical, mediolateral, and anteroposterior directions, respectively.}

\item{\code{lightStartFrame}}{A numeric / integer value depicting the frame in which the light is triggered on the high-speed video.}

\item{\code{startFrame}}{A numeric / integer value depicting the frame number in which the behavior started (e.g., first contact of the foot onto the force plate)}

\item{\code{endFrame}}{A numeric / integer value depicting the frame number in which the behavior ended (e.g., penultimate frame to the foot lifting off from the force plate)}

\item{\code{videoHz}}{A numeric / integer value depicting the frame rate for the high-speed videos.}

\item{\code{forceHz}}{A numeric / integer value depicting the recording rate for the force plate data.}

\item{\code{zeroStart}}{A numeric / integer value depicting the sweep number in which to start zeroing the data (i.e., estimate the baseline). Should be at least 1000 sweeps away from any activity.}

\item{\code{saveData}}{Option to save all of the data from this function as .csv files. Default is set to "no".}
}
\description{
Takes voltage data from a multi-axis force platform and then converts those data to ground reation forces.
}
\details{
These procedures follow the methodology used in Kawano and Blob (2013) and Kawano et al. 2016. It is assumed that the output from the force platform contain 12 channels in the following order: trigger, four verticals, sum of the verticals, two mediolateral, sum of the mediolaterals, two anteroposterior, and the sum of the anteroposteriors.
}
\examples{

voltToForce(calib, lightStartFrame, startFrame, endFrame, videoHz, forceHz, zeroStart, saveData)

}
\references{
Kawano SM, Blob RW. 2013. Propulsive forces of mudskipper fins and salamander limbs during terrestrial locomotion: implications for the invasion of land. Integrative and Comparative Biology 53(2): 283-294. \url{https://academic.oup.com/icb/article/53/2/283/806410/Propulsive-Forces-of-Mudskipper-Fins-and}

Kawano SM, Economy DR, Kennedy MS, Dean D, Blob RW. 2016. Comparative limb bone loading in the humerus and femur of the tiger salamander Ambystoma tigrinum: testing the "mixed-chain" hypothesis for skeletal safety factors. Journal of Experimental Biology 219: 341-353. \url{http://jeb.biologists.org/content/219/3/341}
}
