% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomechanicsFunctions.R
\name{yank}
\alias{yank}
\title{Calculating yank for force data}
\usage{
yank(time, force ...)
}
\arguments{
\item{\code{time}}{Numeric values representing the time elapsed in units of seconds.}

\item{\code{force}}{Numeric values representing the force measured at each time point.}
}
\description{
Calculates yank as a measure of the change in force per unit of time.
}
\details{
These procedures follow the methodology described in Lin et al. 2019.
}
\examples{

yank(seq(1:10), rnorm(10, 1, 1), )

}
\references{
Lin DC, McGowan CP, Blum KP, Ting LH. Yank: the time derivative of force is an important biomechanical variable in sensorimotor systems. Journal of Experimental Biology 222: jeb180414. \url{https://jeb.biologists.org/content/222/18/jeb180414}
}
