% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{Ellipse Area Distribution}
\usage{
area(data)
}
\arguments{
\item{data}{named list of 3 data frames from `breach` function:
`surveyData`, `breachPositionEnsemble` and `breachPositionBestFit`.}
}
\value{
Named list of objects: `ellipse` (sf object of the area of ellipse),
  `fifthPercentileArea` named list containing numeric value in metres of 5%
  percentile area, package version and date, `spotfire_ellipse` (TIBCO
  Spotfire compatible geometry of ellipse), and `polygon` is an sf object
  connecting distance to good points with straight lines (not ellipse).
}
\description{
Calculate the ellipse mixing area distribution and create a ellipse polygon.
}
\examples{
\dontrun{
probability <- probability_non_linear(demo_iqi)
breach <- breach(probability)
area <- area(breach)
plot(area[["ellipse"]])
area[["fifthPercentileArea"]]
}
}
