% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{Assess survey data}
\usage{
assess(
  data,
  overrideTransect1 = NA,
  overrideTransect2 = NA,
  overrideTransect3 = NA,
  overrideTransect4 = NA,
  overrideBearing1 = NA,
  overrideBearing2 = NA,
  overrideBearing3 = NA,
  overrideBearing4 = NA,
  loess = FALSE,
  use_mean_bearing = TRUE,
  ellipse_representative = TRUE
)
}
\arguments{
\item{data}{Data frame with 7 case-sensitive named variables as described
  below, see `demo_iqi` for example data. Rows and columns can be arranged in
  any order. Additional columns can be present but will be ignored.
\describe{
\item{Survey_date}{Survey_date character string}
\item{MCFF}{MCFF Site name - character string}
\item{Transect}{Transect integer}
\item{Station}{Station integer increasing from cage edge e.g. 1,2,3,4 etc}
\item{Easting}{Easting coordinate}
\item{Northing}{Northing coordinate}
\item{IQI}{IQI ratio - Environmental Quality Ratio EQR}
}}

\item{overrideTransect1}{overrideTransect1}

\item{overrideTransect2}{overrideTransect2}

\item{overrideTransect3}{overrideTransect3}

\item{overrideTransect4}{overrideTransect4}

\item{overrideBearing1}{overrideBearing1}

\item{overrideBearing2}{overrideBearing2}

\item{overrideBearing3}{overrideBearing3}

\item{overrideBearing4}{overrideBearing4}

\item{loess}{Use loess model instead of fitting multiple different models}

\item{use_mean_bearing}{Use bearing between the pen edge stations and the
mean coordinate of stations excluding the pen edge.}

\item{ellipse_representative}{Calculate the ellipse based on distance to 95
percentile of resamples on each passing status on each transect (TRUE)
rather than 95 percentile of areas predicted ellipse area resamples
(FALSE). Setting this to TRUE provides an ellipse polygon that reflects the
95% area m2 exactly. Otherwise, the ellipse polygon does not represent
exactly the area in m2.}
}
\value{
A named list containing the fifth percentile of the modeled area in
  meters, package version and package date.
}
\description{
Assess the size of the mixing zone based on modelling the distance to good
status. Where distance cannot be modeled, the distance to the second
consecutive good status (or high status) station is used.
}
\examples{
\dontrun{
area <- assess(demo_iqi)
}
}
