% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach.R
\name{breach}
\alias{breach}
\title{Breach}
\usage{
breach(data, ellipse_representative = FALSE)
}
\arguments{
\item{data}{named list of 4 data frames created by the
`probability_non_linear` function. These are:  `data` (survey), `geoDf`
(distances to good),`geoDfBestFit` (best fit distance) to good and
`hexdfOut` (hexagon heat map).}
}
\value{
Named list containing 3 dataframes: `surveyData`,
  `breachPositionEnsemble` and `breachPositionBestFit`
}
\description{
Calculates the breach distance from centre of ellipse for each transect.
}
\examples{
\dontrun{
probability <- probability_non_linear(demo_iqi)
breach <- breach(probability)
}
}
