% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-hera.R
\name{convert_hera}
\alias{convert_hera}
\title{Convert to hera}
\usage{
convert_hera(method, data, overrides, breachs, areas)
}
\arguments{
\item{method}{Type of method used to analyse samples either 'iqi' or
'residue'.}

\item{data}{dataframe returned from `consecutive_stations()`}

\item{overrides}{dataframe returned from  `override()`}

\item{breachs}{dataframe returned from  `breach()`}

\item{areas}{dataframe returned from `area()`}
}
\value{
dataframe in 'hera' format
}
\description{
Convert kraken output into 'hera' format which is a long format tibble
dataframe
}
