% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-kraken.R
\name{convert_kraken}
\alias{convert_kraken}
\title{Convert to Kraken}
\usage{
convert_kraken(data)
}
\arguments{
\item{data}{}
}
\value{
dataframe
}
\description{
Converts data input from `survey_import()` into 'standard' kraken data
structure.
}
