% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{kraken}
\alias{kraken}
\title{Calculate Footprint}
\usage{
kraken(
  data,
  overrideTransect1 = NA,
  overrideTransect2 = NA,
  overrideTransect3 = NA,
  overrideTransect4 = NA,
  overrideBearing1 = NA,
  overrideBearing2 = NA,
  overrideBearing3 = NA,
  overrideBearing4 = NA,
  loess = FALSE
)
}
\arguments{
\item{loess}{}
}
\description{
Calculate Footprint
}
\examples{
\dontrun{
probability <- kraken(demo_iqi)
}
}
