% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{kraken}
\alias{kraken}
\title{Calculate Footprint}
\usage{
kraken(
  data,
  overrideTransect1 = NA,
  overrideTransect2 = NA,
  overrideTransect3 = NA,
  overrideTransect4 = NA,
  overrideBearing1 = NA,
  overrideBearing2 = NA,
  overrideBearing3 = NA,
  overrideBearing4 = NA,
  loess = FALSE,
  hera_format = FALSE,
  good_moderate = 0.64,
  method = "iqi"
)
}
\arguments{
\item{data}{Input data frame see `demo_iqi`.}

\item{overrideTransect1}{overrideTransect1}

\item{overrideTransect2}{overrideTransect2}

\item{overrideTransect3}{overrideTransect3}

\item{overrideTransect4}{overrideTransect4}

\item{overrideBearing1}{overrideBearing1}

\item{overrideBearing2}{overrideBearing2}

\item{overrideBearing3}{overrideBearing3}

\item{overrideBearing4}{overrideBearing4}

\item{loess}{Use loess model instead of fitting mulitple different models}

\item{hera_format}{Flag if import data in 'hera' format}

\item{good_moderate}{Good-Moderate boundary value}

\item{method}{Type of method used to analyse samples either 'iqi' or
'residue'.}
}
\value{
Data frame contain 8 variables.
}
\description{
Calculate Footprint
}
\examples{
\dontrun{
probability <- kraken(demo_iqi, loess = TRUE)
}
}
