% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-non-linear.R
\name{probability_non_linear}
\alias{probability_non_linear}
\title{Probability Non-linear}
\usage{
probability_non_linear(
  data,
  loess = FALSE,
  pass_fail = 0.64,
  method = "iqi",
  n_try = 1000
)
}
\arguments{
\item{data}{The named data frame `survey_data` from `consecutive_station()`
function. See examples.}

\item{loess}{Use loess model (instead of best fit model).}

\item{pass_fail}{Pass-fail boundary value.}

\item{method}{Type of method used to analyse samples, either "iqi" or
"residue".}

\item{n_try}{Number of attempts to fit a model to bootstrap resamples. The
model is fitted to a maximum 50% of attempts otherwise the model is not
included. By default, n_try = 1000 and therefore 500 bootstraps are
returned if model is successfully fitted.}
}
\value{
list containing four named data frames: data, geoDf, geoDfBestFit and
  hexdfOut.
\describe{
 \item{data}{Summary output of the models fitted for each transect}
 \item{geoDf}{All resample predictions for distance to passing status}
 \item{geoDfBestFit}{Best fit prediction at 95% confidence level for
 distance to passing status based on fitting the model a single time}
 \item{hexdfOut.}{hexagon heat map of resample predictions}
 }
}
\description{
Fit the best model to calculate the distance from pen edge to where the
correct standard is reached. Where distance cannot be modeled, the distance
to the second consecutive station passing the correct standard (or high
status) station is used. If there are not two consecutive stations reaching
the standard, use the distance to the last station (station number 7 or
greater) reaching the correct standard.
}
\examples{
\dontrun{
data <- consecutive_stations(data)
probability <- probability_non_linear(data$survey_data)
}
}
